docs: freeze mart-bill api contract v1

+++ backend/app/db/models/inventory_txn.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import relationship
 
 from .base_class import Base
@@ -13,9 +13,9 @@ class InventoryTxn(Base):
     item_id = Column(Integer, ForeignKey("item.id"), nullable=False)
     batch_id = Column(Integer, ForeignKey("batch.id"), nullable=True)
     txn_type = Column(String(16), nullable=False)  # IN, OUT, ADJUST, etc.
-    raw_qty = Column(Float, nullable=False)
+    raw_qty = Column(Numeric(10, 3), nullable=False)
     raw_unit = Column(String(16), nullable=False)
-    base_qty = Column(Float, nullable=False)
+    base_qty = Column(Numeric(10, 3), nullable=False)
     base_unit = Column(String(16), nullable=False)
     ref_type = Column(String(32), nullable=True)  # 'stock_entry', 'invoice_item', etc.
     ref_id = Column(Integer, nullable=True)
